fix some features

+++ app/src/main/java/com/shoplex/shoplex/model/adapter/CartAdapter.kt
@@ -1,14 +1,20 @@
 package com.shoplex.shoplex.model.adapter
 
 import android.view.LayoutInflater
+import android.view.View
 import android.view.ViewGroup
 import android.widget.Toast
+import androidx.appcompat.app.AlertDialog
+import androidx.coordinatorlayout.widget.CoordinatorLayout
+import androidx.core.content.ContextCompat
 import androidx.recyclerview.widget.RecyclerView
 import com.bumptech.glide.Glide
+import com.google.android.material.snackbar.Snackbar
 import com.shoplex.shoplex.R
 import com.shoplex.shoplex.databinding.RvCartHomeBinding
-import com.shoplex.shoplex.model.pojo.ProductCart
 import com.shoplex.shoplex.model.interfaces.FavouriteCartListener
+import com.shoplex.shoplex.model.pojo.ProductCart
+
 
 class CartAdapter(
     var favouriteCartListener: FavouriteCartListener
@@ -38,7 +44,23 @@ class CartAdapter(
 
             binding.product = product
             binding.imgDelete.setOnClickListener {
-                favouriteCartListener.onDeleteFromCart(product.productID)
+                val builder = binding.root.context?.let { AlertDialog.Builder(it) }
+                builder?.setTitle(binding.root.context.getString(R.string.delete))
+                builder?.setMessage(binding.root.context.getString(R.string.deleteMessage))
+
+                builder?.setPositiveButton(binding.root.context.getString(R.string.yes)) { _, _ ->
+                    favouriteCartListener.onDeleteFromCart(product.productID)
+                    val snackbar = Snackbar.make(binding.root, binding.root.context.getString(R.string.Success), Snackbar.LENGTH_LONG)
+                    val sbView: View = snackbar.view
+                    sbView.setBackgroundColor(ContextCompat.getColor(binding.root.context, R.color.blueshop))
+                    snackbar.show()
+
+                }
+                builder?.setNegativeButton(binding.root.context.getString(R.string.no)) { dialog, _ ->
+                    dialog.cancel()
+                }
+
+                builder?.show()
             }
 
             binding.btnMinus.setOnClickListener {
@@ -58,7 +80,10 @@ class CartAdapter(
                         product.cartQuantity
                     )
                 } else {
-                    Toast.makeText(binding.root.context, "Max Quantity", Toast.LENGTH_SHORT).show()
+                    val snackbar = Snackbar.make(binding.root, binding.root.context.getString(R.string.max), Snackbar.LENGTH_LONG)
+                    val sbView: View = snackbar.view
+                    sbView.setBackgroundColor(ContextCompat.getColor(binding.root.context, R.color.blueshop))
+                    snackbar.show()
                 }
             }
         }
